Name Game: Gender Prediction Using Sound
This is a DataCamp project where we analyze the author names of children's books along with baby names. By using libraries like fuzzy, we can obtain pronunciation sounds of different names. We can then utilize this data to predict the gender of a book's author and the gender distribution in general.