Name Game: Gender Prediction Using Sound
This is a DataCamp project. We analyze author names for different children's books along with several baby names. Using Python libraries like Fuzzy, we can obtain the pronunciation sounds for different names. We can then utilize this data to predict the gender of a book's author or the gender distribution in general.